[AS7-6128] Transformers for Infinispan subsystem VIRTUAL_NODES, SEGMENTS and INDEXING_PROPERTIES.
