+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanExtension.java
@@ -25,11 +25,7 @@ import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.UNDEFINE_ATTRIBUTE_OPERATION;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.ResourceBundle;
 
 import org.jboss.as.controller.Extension;
 import org.jboss.as.controller.ExtensionContext;
@@ -38,7 +34,6 @@ import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.SubsystemRegistration;
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
 import org.jboss.as.controller.descriptions.ResourceDescriptionResolver;
-import org.jboss.as.controller.descriptions.StandardResourceDescriptionResolver;
 import org.jboss.as.controller.parsing.ExtensionParsingContext;
 import org.jboss.as.controller.services.path.ResolvePathHandler;
 import org.jboss.as.controller.transform.DiscardAttributesTransformer;
@@ -129,35 +124,46 @@ public class InfinispanExtension implements Extension {
      */
     private static void registerTransformers(final SubsystemRegistration subsystem) {
 
-        // define the resource and operation transformers
+        // reject expression values for resource - recursive, so have to define all attributes
+        // reject expression values for operations - not recursive, do don't have to define all attributes
 
-        // Register the model transformers
-        TransformersSubRegistration registration = subsystem.registerModelTransformers(ModelVersion.create(1, 3), ResourceTransformer.DEFAULT);
+        InfinispanResourceAndOperationTransformer_1_3 resourceAndOperationTransformer = new InfinispanResourceAndOperationTransformer_1_3() ;
+        final RejectExpressionValuesTransformer totalReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_TOTAL);
+        // shold we chain a resource transformer here to ignore expressions on attributes?
+        final ChainedResourceTransformer chainedResourceAndOperationTransformer = new ChainedResourceTransformer(resourceAndOperationTransformer, totalReject.getChainedTransformer());
 
+        TransformersSubRegistration registration = subsystem.registerModelTransformers(ModelVersion.create(1, 3), chainedResourceAndOperationTransformer);
+
+        // cache-container=*
+        // this transformer will check and reject values for cache-container attributes which should not accept expressions in 1.3
         final RejectExpressionValuesTransformer cacheContainerReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_CONTAINER_ATTRIBUTES);
         TransformersSubRegistration containerRegistration =
                 registerTransformer(registration, CacheContainerResource.CONTAINER_PATH, cacheContainerReject, cacheContainerReject, cacheContainerReject.getWriteAttributeTransformer(), null);
 
+        // cache-container=*/transport=TRANSPORT
+        // this transformer will check and reject values for attributes which should not accept expressions in 1.3
         final RejectExpressionValuesTransformer transportReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_TRANSPORT_ATTRIBUTES);
         registerTransformer(containerRegistration, TransportResource.TRANSPORT_PATH, transportReject, transportReject, transportReject.getWriteAttributeTransformer(), null);
 
-        final InfinispanDiscardAttributesTransformer removeSelectedCacheAttributes = new InfinispanDiscardAttributesTransformer(ModelKeys.INDEXING_PROPERTIES, ModelKeys.SEGMENTS, ModelKeys.VIRTUAL_NODES);
+        // cache-container=*/cache=*
+        // this chained transformer will do two things:
+        // - discard attributes INDEXING_PROPERTIES, SEGMENTS and VIRTUAL_NODES from add and write operations in 1.3
+        // - check and reject values for cache attributes which should not accept expressions in 1.3
+        final InfinispanDiscardAttributesTransformer removeSelectedCacheAttributes = new InfinispanDiscardAttributesTransformer(ModelKeys.INDEXING, ModelKeys.SEGMENTS, ModelKeys.VIRTUAL_NODES);
         final RejectExpressionValuesTransformer cacheReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_CACHE_ATTRIBUTES);
-        final ChainedResourceTransformer chainedResource = new ChainedResourceTransformer(removeSelectedCacheAttributes, cacheReject.getChainedTransformer());
-        final ChainedOperationTransformer chainedAdd = new ChainedOperationTransformer(removeSelectedCacheAttributes, cacheReject);
-        final ChainedOperationTransformer chainedWrite = new ChainedOperationTransformer(removeSelectedCacheAttributes.getWriteAttributeTransformer(), cacheReject.getWriteAttributeTransformer());
-
+        final ChainedResourceTransformer chainedResource = new ChainedResourceTransformer(resourceAndOperationTransformer, cacheReject.getChainedTransformer());
+        final ChainedOperationTransformer chainedAdd = new ChainedOperationTransformer(resourceAndOperationTransformer, cacheReject);
+        final ChainedOperationTransformer chainedWrite = new ChainedOperationTransformer(resourceAndOperationTransformer.getWriteAttributeTransformer(), cacheReject.getWriteAttributeTransformer());
 
         PathElement[] cachePaths = {
                 LocalCacheResource.LOCAL_CACHE_PATH,
                 InvalidationCacheResource.INVALIDATION_CACHE_PATH,
                 ReplicatedCacheResource.REPLICATED_CACHE_PATH,
-                DistributedCacheResource.DISTRIBUTED_CACHE_PATH};
+                DistributedCacheResource.DISTRIBUTED_CACHE_PATH
+        };
         for (int i=0; i < cachePaths.length; i++) {
-            // register chained operation transformers for cache ADD operations where we need to remove and reject
-            TransformersSubRegistration cacheRegistration = //containerRegistration.registerSubResource(cachePaths[i], (OperationTransformer) chainedAdd);
+            TransformersSubRegistration cacheRegistration =
                     registerTransformer(containerRegistration, cachePaths[i], chainedResource, chainedAdd, chainedWrite, removeSelectedCacheAttributes.getUndefineAttributeTransformer());
-
             registerCacheChildrenTransformers(cacheRegistration) ;
         }
     }
@@ -175,6 +181,7 @@ public class InfinispanExtension implements Extension {
 
     private static void registerCacheChildrenTransformers(TransformersSubRegistration cacheReg) {
 
+        // this transformer will check and reject values for cache child attributes which should not accept expressions in 1.3
         final RejectExpressionValuesTransformer childReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_CHILD_ATTRIBUTES) ;
 
         PathElement[] childPaths = {
@@ -190,6 +197,7 @@ public class InfinispanExtension implements Extension {
             cacheReg.registerSubResource(childPaths[i], (OperationTransformer) childReject);
         }
 
+        // this transformer will check and reject values for store attributes which should not accept expressions in 1.3
         final RejectExpressionValuesTransformer storeReject = new RejectExpressionValuesTransformer(InfinispanRejectedExpressions_1_3.REJECT_STORE_ATTRIBUTES);
         PathElement[] storePaths = {
                 StoreResource.STORE_PATH,
@@ -208,139 +216,8 @@ public class InfinispanExtension implements Extension {
         }
     }
 
-    private static class InfinispanResourceDescriptionResolver extends StandardResourceDescriptionResolver {
-
-        // a static map, mapping attribute names to the prefixes that should be used to look up their descriptions
-        private static final Map<String, String> sharedAttributeResolver = new HashMap<String, String>() ;
-
-        static {
-            // shared cache attributes
-            sharedAttributeResolver.put(CacheResource.BATCHING.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.CACHE_MODULE.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.INDEXING.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.INDEXING_PROPERTIES.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.JNDI_NAME.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.NAME.getName(), "cache");
-            sharedAttributeResolver.put(CacheResource.START.getName(), "cache");
-
-            sharedAttributeResolver.put(ClusteredCacheResource.ASYNC_MARSHALLING.getName(), "clustered-cache");
-            sharedAttributeResolver.put(ClusteredCacheResource.MODE.getName(), "clustered-cache");
-            sharedAttributeResolver.put(ClusteredCacheResource.QUEUE_FLUSH_INTERVAL.getName(), "clustered-cache");
-            sharedAttributeResolver.put(ClusteredCacheResource.QUEUE_SIZE.getName(), "clustered-cache");
-            sharedAttributeResolver.put(ClusteredCacheResource.REMOTE_TIMEOUT.getName(), "clustered-cache");
-
-            sharedAttributeResolver.put(BaseStoreResource.FETCH_STATE.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.PASSIVATION.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.PRELOAD.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.PURGE.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.SHARED.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.SINGLETON.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.PROPERTY.getName(), "store");
-            sharedAttributeResolver.put(BaseStoreResource.PROPERTIES.getName(), "store");
-
-            sharedAttributeResolver.put(BaseJDBCStoreResource.DATA_SOURCE.getName(), "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.BATCH_SIZE.getName(), "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.FETCH_SIZE.getName(), "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.PREFIX.getName(), "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.ID_COLUMN.getName()+".column", "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.DATA_COLUMN.getName()+".column", "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.TIMESTAMP_COLUMN.getName()+".column", "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.ENTRY_TABLE.getName()+"table", "jdbc-store");
-            sharedAttributeResolver.put(BaseJDBCStoreResource.BUCKET_TABLE.getName()+"table", "jdbc-store");
-
-            // shared children - this avoids having to describe the children for each parent resource
-            sharedAttributeResolver.put(ModelKeys.TRANSPORT, null);
-            sharedAttributeResolver.put(ModelKeys.LOCKING, null);
-            sharedAttributeResolver.put(ModelKeys.TRANSACTION, null);
-            sharedAttributeResolver.put(ModelKeys.EVICTION, null);
-            sharedAttributeResolver.put(ModelKeys.EXPIRATION, null);
-            sharedAttributeResolver.put(ModelKeys.STATE_TRANSFER, null);
-            sharedAttributeResolver.put(ModelKeys.STORE, null);
-            sharedAttributeResolver.put(ModelKeys.FILE_STORE, null);
-            sharedAttributeResolver.put(ModelKeys.REMOTE_STORE, null);
-            sharedAttributeResolver.put(ModelKeys.STRING_KEYED_JDBC_STORE, null);
-            sharedAttributeResolver.put(ModelKeys.BINARY_KEYED_JDBC_STORE, null);
-            sharedAttributeResolver.put(ModelKeys.MIXED_KEYED_JDBC_STORE, null);
-            sharedAttributeResolver.put(ModelKeys.WRITE_BEHIND, null);
-            sharedAttributeResolver.put(ModelKeys.PROPERTY, null);
-        }
-
-        private InfinispanResourceDescriptionResolver(String keyPrefix, String bundleBaseName, ClassLoader bundleLoader) {
-            super(keyPrefix, bundleBaseName, bundleLoader, true, false);
-        }
-
-        @Override
-        public String getResourceAttributeDescription(String attributeName, Locale locale, ResourceBundle bundle) {
-            // don't apply the default bundle prefix to these attributes
-            if (sharedAttributeResolver.containsKey(attributeName)) {
-               return bundle.getString(getBundleKey(attributeName));
-            }
-            return super.getResourceAttributeDescription(attributeName, locale, bundle);
-        }
-
-        @Override
-        public String getResourceAttributeValueTypeDescription(String attributeName, Locale locale, ResourceBundle bundle, String... suffixes) {
-            // don't apply the default bundle prefix to these attributes
-            if (sharedAttributeResolver.containsKey(attributeName)) {
-               return bundle.getString(getVariableBundleKey(attributeName, suffixes));
-            }
-            return super.getResourceAttributeValueTypeDescription(attributeName, locale, bundle, suffixes);
-        }
-
-        @Override
-        public String getOperationParameterDescription(String operationName, String paramName, Locale locale, ResourceBundle bundle) {
-            // don't apply the default bundle prefix to these attributes
-            if (sharedAttributeResolver.containsKey(paramName)) {
-               return bundle.getString(getBundleKey(paramName));
-            }
-            return super.getOperationParameterDescription(operationName, paramName, locale, bundle);
-        }
-
-        @Override
-        public String getOperationParameterValueTypeDescription(String operationName, String paramName, Locale locale, ResourceBundle bundle, String... suffixes) {
-            // don't apply the default bundle prefix to these attributes
-            if (sharedAttributeResolver.containsKey(paramName)) {
-               return bundle.getString(getVariableBundleKey(paramName, suffixes));
-            }
-            return super.getOperationParameterValueTypeDescription(operationName, paramName, locale, bundle, suffixes);
-        }
-
-        @Override
-        public String getChildTypeDescription(String childType, Locale locale, ResourceBundle bundle) {
-            // don't apply the default bundle prefix to these attributes
-            if (sharedAttributeResolver.containsKey(childType)) {
-               return bundle.getString(getBundleKey(childType));
-            }
-            return super.getChildTypeDescription(childType, locale, bundle);
-        }
-
-        private String getBundleKey(final String name) {
-            return getVariableBundleKey(name);
-        }
-
-        private String getVariableBundleKey(final String name, final String... variable) {
-            final String prefix = sharedAttributeResolver.get(name);
-            StringBuilder sb = new StringBuilder(SUBSYSTEM_NAME);
-            // construct the key prefix
-            if (prefix == null) {
-                sb = sb.append('.').append(name);
-            } else {
-                sb = sb.append('.').append(prefix).append('.').append(name);
-            }
-            // construct the key suffix
-            if (variable != null) {
-                for (String arg : variable) {
-                    if (sb.length() > 0)
-                        sb.append('.');
-                    sb.append(arg);
-                }
-            }
-            return sb.toString();
-        }
-    }
-
     private static class InfinispanDiscardAttributesTransformer extends DiscardAttributesTransformer {
-        InfinispanDiscardAttributesTransformer(String...attributes){
+        private InfinispanDiscardAttributesTransformer(String... attributes) {
             super(attributes);
         }
     }
