+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanRejectedExpressions_1_3.java
@@ -126,6 +126,14 @@ public class InfinispanRejectedExpressions_1_3 {
             ACCEPT14_ACCEPT13_STORE_ATTRIBUTES
     );
 
+    public static final AttributeDefinition[] REJECT_TOTAL =
+            concat(
+                REJECT_CONTAINER_ATTRIBUTES,
+                REJECT_TRANSPORT_ATTRIBUTES,
+                REJECT_CACHE_ATTRIBUTES,
+                REJECT_CHILD_ATTRIBUTES,
+                REJECT_STORE_ATTRIBUTES
+            ) ;
 
     // attributes which accept in 1.4.0 but reject in 1.3.0 only
     // set = all - never accept - always accept
