+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanResourceAndOperationTransformer_1_3.java
@@ -23,24 +23,35 @@
 package org.jboss.as.clustering.infinispan.subsystem;
 
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.NAME;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.VALUE;
 
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.registry.Resource;
 import org.jboss.as.controller.transform.AbstractOperationTransformer;
+import org.jboss.as.controller.transform.OperationResultTransformer;
 import org.jboss.as.controller.transform.OperationTransformer;
 import org.jboss.as.controller.transform.TransformationContext;
 import org.jboss.as.controller.transform.chained.ChainedResourceTransformationContext;
 import org.jboss.as.controller.transform.chained.ChainedResourceTransformerEntry;
 import org.jboss.dmr.ModelNode;
+import org.jboss.dmr.ModelType;
 
 /**
+ * Convert the 1.4 SEGMENTS value to VIRTUAL_NODES in model and operations, if defined and not an expression
+ * Remove the 1.4 attributes INDEXING_PROPERTIES and SEGMENTS from model and operations
+ *
  * @author <a href="mailto:tomaz.cerar@redhat.com">Tomaz Cerar</a>
+ * @author Richard Achmatowicz (c) RedHat 2013
  */
 public class InfinispanResourceAndOperationTransformer_1_3 extends AbstractOperationTransformer implements ChainedResourceTransformerEntry {
 
+    // ratio of segments to virtual nodes to convert between the two
+    public static final int SEGMENTS_PER_VIRTUAL_NODE = 6;
+
     @Override
     protected ModelNode transform(TransformationContext context, PathAddress address, ModelNode operation) {
+        convert(operation);
         remove(operation);
         return operation;
     }
@@ -48,6 +59,7 @@ public class InfinispanResourceAndOperationTransformer_1_3 extends AbstractOpera
     @Override
     public void transformResource(ChainedResourceTransformationContext context, PathAddress address, Resource resource)
             throws OperationFailedException {
+        convert(resource.getModel());
         remove(resource.getModel());
     }
 
@@ -55,31 +67,77 @@ public class InfinispanResourceAndOperationTransformer_1_3 extends AbstractOpera
         return IGNORE;
     }
 
-    private void remove(ModelNode model) {
-        if (model.has(ModelKeys.INDEXING_PROPERTIES)){
-            model.remove(ModelKeys.INDEXING_PROPERTIES);
-        }
-        if (model.has(ModelKeys.SEGMENTS)) {
-            model.remove(ModelKeys.SEGMENTS);
-        }
-        if (model.has(ModelKeys.VIRTUAL_NODES)) {
-            model.remove(ModelKeys.VIRTUAL_NODES);
-        }
-    }
 
     private static final OperationTransformer IGNORE = new OperationTransformer() {
         @Override
         public TransformedOperation transformOperation(TransformationContext context, PathAddress address, ModelNode operation)
                 throws OperationFailedException {
+
             String attrName = operation.get(NAME).asString();
-            if (attrName.equals(ModelKeys.INDEXING_PROPERTIES) || attrName.equals(ModelKeys.SEGMENTS)) {
+            ModelNode attrValue = operation.get(VALUE);
+
+            if (attrName.equals(ModelKeys.INDEXING_PROPERTIES)) {
                 return OperationTransformer.DISCARD.transformOperation(context, address, operation);
             }
-
+            if (attrName.equals(ModelKeys.SEGMENTS)) {
+                // if the SEGMENTS value is an expression, we can't convert it, so just discard
+                if (!isExpression(attrValue)) {
+                    // transform the operation to a write of virtual nodes
+                    ModelNode transformed = operation.clone();
+                    // convert segments value to virtual-nodes value, and replace SEGMENTS with VIRTUAL_NODES
+                    int segmentsValue = Integer.parseInt(attrValue.asString());
+                    int virtualNodes = (segmentsValue / SEGMENTS_PER_VIRTUAL_NODE) + 1;
+                    transformed.get(NAME).set(ModelKeys.VIRTUAL_NODES);
+                    transformed.get(VALUE).set(Integer.toString(virtualNodes));
+                    // return the new transformed operation model node;  this is a write, so the return result is unaffected
+                    return new TransformedOperation(transformed, OperationResultTransformer.ORIGINAL_RESULT);
+                }  else {
+                    return OperationTransformer.DISCARD.transformOperation(context, address, operation);
+                }
+            }
             return OperationTransformer.DEFAULT.transformOperation(context, address, operation);
         }
     };
 
+    private static boolean isExpression(ModelNode node) {
+        return node.getType() == ModelType.EXPRESSION ;
+    }
 
+    private static void remove(ModelNode model) {
+        if (model.has(ModelKeys.INDEXING_PROPERTIES)){
+            model.remove(ModelKeys.INDEXING_PROPERTIES);
+        }
+        if (model.has(ModelKeys.SEGMENTS)){
+            model.remove(ModelKeys.SEGMENTS);
+        }
+    }
 
+    private static void convert(ModelNode model) {
+        if (model.has(ModelKeys.SEGMENTS)) {
+            ModelNode segments = model.get(ModelKeys.SEGMENTS) ;
+            if (segments.isDefined() && !isExpression(segments)) {
+                // convert segments value to virtual-nodes value
+                int segmentsValue = Integer.parseInt(segments.asString());
+                int virtualNodes = (segmentsValue / SEGMENTS_PER_VIRTUAL_NODE) + 1;
+                model.get(ModelKeys.VIRTUAL_NODES).set(Integer.toString(virtualNodes));
+            }
+            model.remove(ModelKeys.SEGMENTS);
+        }
+    }
+
+    /*
+     * Convert a 1.3 virtual nodes value to a 1.4 segments value
+     * TODO: handle expressions
+     */
+    public static String virtualNodesToSegments(String virtualNodesValue) {
+        int segments = 0 ;
+        try {
+            segments =  Integer.parseInt(virtualNodesValue) * SEGMENTS_PER_VIRTUAL_NODE;
+        }
+        catch(NumberFormatException nfe) {
+            // in case of expression
+        }
+        return Integer.toString(segments);
+    }
 }
+
