+++ clustering/infinispan/src/main/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanSubsystemXMLReader_1_3.java
@@ -22,7 +22,6 @@
 
 package org.jboss.as.clustering.infinispan.subsystem;
 
-import static org.jboss.as.clustering.infinispan.InfinispanLogger.ROOT_LOGGER;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
 
@@ -322,7 +321,9 @@ public final class InfinispanSubsystemXMLReader_1_3 implements XMLElementReader<
                     break;
                 }
                 case VIRTUAL_NODES: {
-                    ROOT_LOGGER.virtualNodesAttributeDeprecated();
+                    // AS7-5753: convert any non-expression virtual nodes value to a segments value,
+                    DistributedCacheResource.SEGMENTS.parseAndSetParameter(
+                            InfinispanResourceAndOperationTransformer_1_3.virtualNodesToSegments(value), cache, reader);
                     break;
                 }
                 case L1_LIFESPAN: {
